enable control_mask, delete update_ops_quantum_matrix.cpp
